Refactor openssl_csr module into backend + module.
